requirements.txt
opentelemetry-api == 1.38.0
opentelemetry-sdk == 1.38.0
opentelemetry-exporter-otlp-proto-grpc == 1.38.0
opentelemetry-semantic-conventions